refactor: Moved userApps file to helpers
