+++ src/sdk/helpers/userApps.ts
@@ -3,10 +3,10 @@ import { Name, PublicKey } from '@wharfkit/antelope';
 import { IDContract } from '../services/blockchain/contracts/IDContract';
 import { KeyManager, KeyManagerLevel } from '../storage/keymanager';
 import { createStorage, PersistentStorageClean, StorageFactory, STORAGE_NAMESPACE } from '../storage/storage';
-import { User } from './user';
+import { User } from '../controllers/user';
 import { createKeyManagerSigner } from '../services/blockchain/eosio/transaction';
 import { SdkError, SdkErrors, throwError } from '../util/errors';
-import { App, AppStatus } from './app';
+import { App, AppStatus } from '../controllers/app';
 import { TonomyUsername } from '../util/username';
 import { LoginRequest, TonomyRequest } from '../util/request';
 import { LoginRequestResponseMessage, LoginRequestsMessagePayload } from '../services/communication/message';
